picoassist/pyproject.toml
[project]
name = "picoassist"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "msal==1.31.1",
    "httpx==0.28.1",
    "playwright==1.49.1",
    "pydantic==2.10.4",
    "pyyaml==6.0.2",
    "python-dotenv==1.0.1",
    "google-auth>=2.27.0",
    "google-auth-oauthlib>=1.2.0",
    "aiosqlite==0.20.0",
    "psutil>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.25.0",
    "pytest-mock==3.14.0",
    "respx==0.22.0",
    "ruff==0.8.6",
]

[tool.setuptools.packages.find]
include = ["services*", "config*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["services", "tests", "config"]
markers = ["live: requires running PicoAssist services on localhost"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
